feat(client): Refactor restructure StatisticsPage components + Improve useChartWrapperSize (to trigger re-render on tab switch) + Remove PieChart
