--- client/src/app/components/stardust/statistics/charts/PieChart.scss
@@ -1,5 +1,0 @@
-.pie-chart-wrapper {
-    .pie-chart-wrapper__content {
-        height: 480px;
-    }
-}
